Database Update These Changes are Not yet fetlected in code and there will be errors!
Booking
 CheckedInDate - added property
 CheckedOutDate - Added Property
 BookingStatus - added Fk Table
 VolumeAdults changed to correct spelling

BookingStatus - new table
 id - added property
 bookingstatusdescription - added property

Building
 changed building1 to buildingname
 changed phone from double to int64

maintenanceLog
 changed date from string to date\

Room
 room number changed to string from int to allow for appt letters

RoomType
 Bathrooms changed from int to Decimal

HouseKeepingstatu
 changed table name to HouseKeepingstatus

Customer
 phone changed to int64 from double

Expense
 Table Renamed to Expenses

Employee
 phone changed from double to int64

CreatedEmailTable
 Id int32
 body - string
 DateSent - DateTime
 EmployeeId - Fk Employee Table

EmailRecipients
 Id - int32
 EmailAddress - string
 EmaisId - Fk to Emails table

+++ cmps285HotelProject/HotelIntegratedComputerSystems/HotelIntegratedComputerSystems/Models/Customer.cs
@@ -23,7 +23,7 @@ namespace HotelIntegratedComputerSystems.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
-        public double Phone { get; set; }
+        public long Phone { get; set; }
         public string Email { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
